requirements.txt
click==6.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
psycopg2==2.6.1
PyJWT==1.4.0
SQLAlchemy==1.0.13
Werkzeug==0.11.10
